last of the type signature unfuckery i think

+++ lab8/buffer.c
@@ -10,20 +10,20 @@
 #include <sys/stat.h> /* For mode constants */
 #include <unistd.h>
 
-SBuffer* setupBuffer(char* zone, int nbTasks)
+SBuffer* setupBuffer(char* zone, size_t nbTasks)
 {
     // Create and initialize a memory mapping in the zone named "zone" capable
     // of holding nbTasks. return the address of the shared buffer.
 }
 
-SBuffer* makeBuffer(void* z, int sz, size_t mapSize)
+SBuffer* makeBuffer(void* z, size_t sz, size_t mapSize)
 {
     // Given a memory mapping at address z, meant to hold sz Task for a total
     // size (in bytes) of mapSize initialize and return a shared buffer pointer
     // held in that mapping.
 }
 
-SBuffer* getBuffer(char* zone, int nbTasks)
+SBuffer* getBuffer(char* zone, size_t nbTasks)
 {
     // Given a named memory mapping zone (capable of holding nbTasks), return
     // the address of the shared buffer held at the beginning of that memory
